requirements.txt
aiohttp~=3.7.4post0
aiogram~=2.14.3
psycopg2-binary~=2.9.1
boto~=2.49.0